Revert "further move models where they belong"
This reverts commit 50c20facaf274f936d0a8ad86012820d8c9b965e.
