+++ backend/auth0login/views.py
@@ -34,7 +34,7 @@ def dashboard(request):
         userdata = {
             'name': user.first_name,
             'alpaca_id': alpaca_id,
-            'alpaca_key': "*" * len(alpaca_key),
+            'alpaca_key': "*"*len(alpaca_key),
             'total_equity': user_details['equity'],
             'buy_power': user_details['buy_power'],
             'portfolio': user_details['portfolio'],
@@ -122,6 +122,7 @@ def sync_alpaca(user):
     # 3) check if user has stock instance
     # for position in portfolio:
     #
+
 
     user_details['portfolio'] = portfolio
     return user_details
